Build: Tentando realizar build utilizando gitactions para aws

+++ .docker/start-back.sh
@@ -6,11 +6,19 @@ composer install
 
 chmod +x /var/www/html/artisan
 
-if [ ! -f ".env" ]; then
+if [ ! -f ".env" ] && [ "$APP_ENV" == "production" ]; then
+    cp .env.prod .env
+else 
     cp .env.example .env
 fi
 
-php artisan migrate --force
+if [ "$APP_ENV" == "production" ]; then
+    echo "Executando as migrations em produção"
+    php artisan migrate --force
+else
+    echo "Executando as migrations local"
+    php artisan migrate --force
+fi
 
 php artisan route:cache &&
 php artisan config:cache &&
